+++ include/stats.h
@@ -5,6 +5,6 @@
 #define MAX_DATASET_SIZE 100
 
 int compare(const void *a, const void *b); // Array Sorting function for qsort
-double stats(float arr[], int n, int subChoice); // Statistics functions
+double stats(double arr[], int n, int subChoice); // Statistics functions
 
 #endif
